+++ db/migrations/0005_channel_credentials_encrypted.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "channel" ADD COLUMN "credentials_encrypted" text;